requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
mysqlclient==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
flask-cors==3.0.10

